+++ packages/transaction-pay-controller/CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
-- Updated TransactionPayBridgeQuote types to account for multichain quotes ([#6862](https://github.com/MetaMask/core/pull/6862))
+- Updated getBridgeBatchTransactions types to account for multichain approvals ([#6862](https://github.com/MetaMask/core/pull/6862))
 
 ## [3.1.0]